Add files via upload
Revised versions, using srand and rand functions to randomly generate marks, also added an option to search for specific data or display it all at once. Added a do-while loop and added a choice to repeat or exit the program. Made changes onto the struct in to use code more efficiently